Board - fix error when logging out - Ref #343

+++ src/main/java/com/shanebeestudios/skbee/elements/scoreboard/objects/Board.java
@@ -89,6 +89,7 @@ public class Board {
     }
 
     public void deleteFastboard() {
+        if (this.fastBoard == null) return;
         this.fastBoard.delete();
     }
